refactor(infra): extract repo discovery and worktree lock
